// models/Enum.ts
export enum HealthStatus {
    Full = "full",
    Healthy = "healthy",
    Poor = "poor",
    Critical = "critical",
}

export enum Category {
    WOODBOARD = "WOODBOARD",
    FITTING = "FITTING",
    ELECTRICAL = "ELECTRICAL",
    EDGE = "EDGE",
    LAMINATE = "LAMINATE",
    PAINTING = "PAINTING",
    ACCESSORIES = "ACCESSORIES",
    STATIONERY = "STATIONERY",
    OTHERS = "OTHERS",
    HANDTOOLS = "HANDTOOLS",
    "POWER TOOLS" = "POWER TOOLS",
}
export enum Type {
    Stock = "Stock",
    Project = "Project",
    Cancel = "Cancel",
}
export enum No {
    WOODBOARD = "WB",
    FITTING = "FT",
    ELECTRICAL = "EE",
    EDGE = "EG",
    LAMINATE = "LM",
    PAINTING = "PA",
    ACCESSORIES = "AC",
    STATIONERY = "SN",
    OTHERS = "OH",
    HANDTOOLS = "HT",
    "POWER TOOLS" = "PT",
}

export enum Unit {
    แผ่น = "แผ่น",
    ถัง = "ถัง",
    ควอท = "ควอท",
    แกลลอน = "แกลลอน",
    ถังกลาง = "ถังกลาง",
    ถังใหญ่ = "ถังใหญ่",
    ชิ้น = "ชิ้น",
    ปี๊ป = "ปี๊ป",
    กล่อง = "กล่อง",
    ม้วน = "ม้วน",
    เส้น = "เส้น",
    ถุง = "ถุง",
    ชุด = "ชุด",
    ตัว = "ตัว",
    แพค = "แพค",
    ดอก = "ดอก",
    มัด = "มัด",
    กิโล = "กิโล",
    หลอด = "หลอด",
    กระสอบ = "กระสอบ",
    แท่ง = "แท่ง",
    กระป๋อง = "กระป๋อง",
    ก้อน = "ก้อน",
}